+++ app/containers/AgentDesktop/tests/selectors.test.js
@@ -9,6 +9,9 @@ import {
   selectAgentDesktopMap,
   selectAgentId,
   selectAwaitingDisposition,
+  selectHasVoiceInteraction,
+  selectSmsInteractionNumbers,
+  selectInteractionEmails,
 } from '../selectors';
 
 describe('selectInteractionsList', () => {
@@ -75,8 +78,101 @@ describe('selectAwaitingDisposition', () => {
           },
         ],
         noInteractionContactPanel: {},
+        newInteractionPanel: {},
       },
     });
     expect(selectAwaitingDisposition(mockedState)).toEqual(true);
+  });
+});
+
+describe('selectHasVoiceInteraction', () => {
+  it('should return true when there is a voice interaction', () => {
+    const mockedState = fromJS({
+      agentDesktop: {
+        interactions: [
+          {
+            channelType: 'sms',
+          },
+          {
+            channelType: 'voice',
+          },
+        ],
+      },
+    });
+    expect(selectHasVoiceInteraction(mockedState)).toEqual(true);
+  });
+  it('should return false when there is not a voice interaction', () => {
+    const mockedState = fromJS({
+      agentDesktop: {
+        interactions: [
+          {
+            channelType: 'sms',
+          },
+        ],
+      },
+    });
+    expect(selectHasVoiceInteraction(mockedState)).toEqual(false);
+  });
+});
+
+describe('selectSmsInteractionNumbers', () => {
+  it('should return the customer numbers for all sms interactions', () => {
+    const mockedState = fromJS({
+      agentDesktop: {
+        interactions: [
+          {
+            channelType: 'email',
+            customer: 'test@serenova.com',
+          },
+          {
+            channelType: 'sms',
+            customer: '+15064701234',
+          },
+          {
+            channelType: 'sms',
+            customer: '+15064709876',
+          },
+          {
+            channelType: 'email',
+            customer: 'test2@serenova.com',
+          },
+        ],
+      },
+    });
+    expect(selectSmsInteractionNumbers(mockedState)).toEqual([
+      '+15064701234',
+      '+15064709876',
+    ]);
+  });
+});
+
+describe('selectInteractionEmails', () => {
+  it('should return the customer emails for all email interactions', () => {
+    const mockedState = fromJS({
+      agentDesktop: {
+        interactions: [
+          {
+            channelType: 'email',
+            customer: 'test@serenova.com',
+          },
+          {
+            channelType: 'sms',
+            customer: '+15064701234',
+          },
+          {
+            channelType: 'sms',
+            customer: '+15064709876',
+          },
+          {
+            channelType: 'email',
+            customer: 'test2@serenova.com',
+          },
+        ],
+      },
+    });
+    expect(selectInteractionEmails(mockedState)).toEqual([
+      'test@serenova.com',
+      'test2@serenova.com',
+    ]);
   });
 });